solucion

+++ Prácticas/P2/Prac2/prac2.cc
@@ -155,7 +155,7 @@ bool Argumentos(int argc, char *argv[], string &inputFile, bool &mostrarEstadist
 int searchPatient(string nif, Database &data)
 {
     int posicion = -1;
-    for (int i = 0; i < (int)data.patients.size(); i++)
+    for (size_t i = 0; i < (int)data.patients.size(); i++)
     {
         if (data.patients[i].nif == nif)
             posicion = i;
@@ -268,13 +268,13 @@ void viewPatient(Database &data)
                 cout << "Telephone: " << paciente.telephone << endl;
                 // Aqui daba error porque ponia 'bool hayAnalysis = false;' y no se usaba
 
-                for (int i = 0; i < data.analysis.size(); i++)
+                for (size_t i = 0; i < data.analysis.size(); i++)
                 {
                     if (strcmp(data.analysis[i].nif, paciente.nif.c_str()) == 0)
                     {
                         // Aqui daba error porque ponia 'hayAnalysis = true;' y no se usaba
                         cout << "id\tdate\theight\tweight" << endl;
-                        for (int j = 0; j < data.analysis.size(); j++)
+                        for (size_t j = 0; j < data.analysis.size(); j++)
                         {
                             cout << data.analysis[j].id << "\t";
                             cout << data.analysis[j].dateAnalysis.day << "/";
@@ -322,7 +322,7 @@ void deletePatient(Database &data)
             }
             else
             {
-                for (int i = 0; data.analysis.size();)
+                for (size_t i = 0; data.analysis.size();)
                 {
                     if (strcmp(data.analysis[i].nif, nif.c_str()))
                     {
@@ -384,7 +384,7 @@ void savePatients(const Database &data)
     ofstream fichero("patients.bin", ios::binary);
     if (fichero.is_open())
     {
-        for (int i = 0; i < data.patients.size(); i++)
+        for (size_t i = 0; i < data.patients.size(); i++)
         {
             PatientBin pacienteBinario;
             strncpy(pacienteBinario.nif, data.patients[i].nif.c_str(), KMAXNIF);
@@ -494,7 +494,7 @@ void exportAnalysis(const Database &data)
     ofstream fichero("analysis.bin", ios::binary);
     if (fichero.is_open())
     {
-        for (int i = 0; i < data.analysis.size(); i++)
+        for (size_t i = 0; i < data.analysis.size(); i++)
         {
             fichero.write((const char *)&data.analysis[i], sizeof(Analysis));
         }
@@ -552,7 +552,7 @@ void Statistics(Database &data)
     float weight;
     float height;
     float IMC;
-    for (int i = 0; data.analysis.size(); i++)
+    for (size_t i = 0; data.analysis.size(); i++)
     {
         day = data.analysis[i].dateAnalysis.day;
         month = data.analysis[i].dateAnalysis.month;
